requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
Pillow==10.0.0
numpy==1.24.3
requests==2.31.0
